% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalia_CE90.R
\name{avalia_ce90}
\alias{avalia_ce90}
\title{Funcao avalia_ce90 -  avalia_ce90 function}
\usage{
avalia_ce90(x, y)
}
\arguments{
\item{x}{entrada de vetor numerico de diferencas planimetricas leste.}

\item{y}{entrada de vetor numerico de diferencas planimetricas norte.}
}
\value{
imprime valor de CE 90 porcento da amostra analisada
}
\description{
Calcula  90 porcento do erro circular CE90 (planimetria) para verificacao do PEC
}
\examples{
set.seed(123)
x <- rnorm(30,1.5,0.5)
y <- rnorm(30,2,1)
avalia_ce90(x,y)

}
