% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalia.R
\name{avalia_le90}
\alias{avalia_le90}
\title{Funcao avalia_le90 - avalia_le90 function}
\usage{
avalia_le90(x)
}
\arguments{
\item{x}{entrada de vetor numerico de diferencas de altitudes.}
}
\value{
imprime valor de le90 da amostra analisada
}
\description{
Calcula 90 porcento do erro linear, le90 em altimetria para verificacao do PEC
}
\examples{
set.seed(123)
x <- rnorm(30,1.5,0.5)
avalia_le90(x)

}
