% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalia.R
\name{avalia_resumo}
\alias{avalia_resumo}
\title{Funcao avalia_resumo - avalia_resumo function}
\usage{
avalia_resumo(x)
}
\arguments{
\item{x}{entrada de vetor numerico, pode ter NAs.}
}
\value{
Imprime resumo com as estatisticas analisadas
}
\description{
Calcula e imprime numero de observacoes, media, desvpad, mediana, min, max, amplitude e NAs.
}
\examples{
set.seed(123)
x <- rnorm(30,1.5,0.5)
avalia_resumo(x)

}
