% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{ce90}
\alias{ce90}
\title{Funcao ce90 -  ce90 function}
\usage{
ce90(x)
}
\arguments{
\item{x}{entrada de vetor numérico pode ter NAs.}
}
\value{
imprime valor de CE 90 porcento da amostra analisada
}
\description{
cálculo del 90 porcento do erro circular CE90 para verificacao do PEC
}
\examples{
ce90(DeltaN)
ce90(DeltaH)
}
