% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{eqm}
\alias{eqm}
\title{Funcao eqm - eqm function}
\usage{
eqm(x)
}
\arguments{
\item{x}{entrada de vetor numerico pode ter NAs.}
}
\value{
imprime valor de EQM da amostra analisada
}
\description{
calculo del Erro Quadratico Medio para verificacao do PEC
}
\examples{
eqm(DeltaH)
eqm(DeltaN)
}
