% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{le90}
\alias{le90}
\title{Funcao le90 - le90 function}
\usage{
le90(x)
}
\arguments{
\item{x}{entrada de vetor numerico pode ter NAs.}
}
\value{
imprime valor de LE90 da amostra analisada
}
\description{
calculo del 90\%  do erro,  LE90 para verificacao do PEC
}
\examples{
le90(DeltaN)
le90(DeltaH)
}
