% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise.R
\name{resumo}
\alias{resumo}
\title{Funcao resumo - resumo function}
\usage{
resumo(x)
}
\arguments{
\item{x}{entrada de vetor numerico, pode ter NAs.}
}
\value{
Imprime resumo com as estatesticas analisadas
}
\description{
Calcula e imprime numero observ., media, desvpad, mediana, min, max, amplitude e NAs.
}
\examples{
resumo (DeltaN)
resumo(DeltaE)
}
