% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalia.R
\name{avalia_eqm}
\alias{avalia_eqm}
\title{Funcao avalia_eqm - avalia_eqm function}
\usage{
avalia_eqm(x)
}
\arguments{
\item{x}{entrada de vetor numerico pode ter NAs.}
}
\value{
imprime valor de EQM da amostra analisada
}
\description{
calcula Erro Quadratico Medio para verificacao do PEC
}
\examples{
set.seed(123)
x <- rnorm(30,1.5,0.5)
avalia_eqm(x)
}
